I’ve re-named the bucketInHostname config param to pathStyleAccess to match the name used in the s3ClientOptions in the AWS java library. Note that this has the opposite sense of the previous flag, i.e. false means put the bucket name in as part of the hostname.
I’ve added unit tests for the S3.url function and to check that the config default is correct for the pathStyleAccess parameter. I’ve re-factored the tests so that the config tests are separated from the bucket tests as a convenience. The bucket tests make actual calls to S3 so it’s hard to run them casually, whilst the config specs to not rely on external resources and can be run quickly and easily.