Remove default prefix and implicit from fromConfiguration
After getting feedback from @EECOLOR, removing the implicit keyword from
the fromConfiguration method since there isn't a need for it.

Also removing the default prefix of "aws", this is done by line 21 in
the non-prefixed version of the fromConfiguration method. So we should
expect people to specify their prefix's when calling for that
functionality specifically.